+++ src/editor/core/plugins/link/schema.js
@@ -5,6 +5,8 @@
  *
  */
 
+import {validateHref} from "../../util/linkUtil";
+
 const schema = {
     marks: {
         sortOrder: 300,
@@ -21,7 +23,7 @@ const schema = {
                 [{
                     tag: "a[href]", getAttrs: function getAttrs(dom) {
                         let href = dom.getAttribute("href");
-                        if (!/^https?:\/\//i.test(href) && !/^mailto:/i.test(href) && !/^ftps?:\/\//i.test(href))  {
+                        if (!validateHref(href))  {
                             href = '#';
                         }
 
@@ -34,9 +36,9 @@ const schema = {
                     }
                 }],
             toDOM: (node) => {
-                let href = (window.humhub && node.attrs.fileGuid) ? humhub.modules.file.getFileUrl(node.attrs.fileGuid)  : node.attrs.href;
+                let href = (window.humhub && node.attrs.fileGuid) ? humhub.modules.file.getFileUrl(node.attrs.fileGuid) : node.attrs.href;
 
-                if (!/^https?:\/\//i.test(href) && !/^mailto:/i.test(href) && !/^ftps?:\/\//i.test(href))  {
+                if (!validateHref(href))  {
                     href = '#';
                 }
 
@@ -53,7 +55,7 @@ const schema = {
                     let href = (window.humhub) ? humhub.modules.file.filterFileUrl(tok.attrGet("href")).url : tok.attrGet("href");
                     let fileGuid = (window.humhub) ? humhub.modules.file.filterFileUrl(tok.attrGet("href")).guid : null;
 
-                    if (!/^https?:\/\//i.test(href) && !/^mailto:/i.test(href) && !/^ftps?:\/\//i.test(href))  {
+                    if (!validateHref(href))  {
                         href = '#';
                     }