+++ src/editor/core/util/linkUtil.js
@@ -1,0 +1,3 @@
+export function validateHref(href) {
+    return /^https?:\/\//i.test(href) || /^mailto:/i.test(href) || /^ftps?:\/\//i.test(href);
+}